HSEARCH-3711 Rename a few misnamed tests
