+++ documentation/src/test/java/org/hibernate/search/documentation/mapper/orm/bridge/propertybridge/parameter/PropertyBridgeParameterIT.java
@@ -4,16 +4,16 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.documentation.mapper.orm.bridge.valuebridge.binder;
+package org.hibernate.search.documentation.mapper.orm.bridge.propertybridge.parameter;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.math.BigDecimal;
 import java.util.List;
 import javax.persistence.EntityManagerFactory;
 
 import org.hibernate.search.documentation.testsupport.BackendConfigurations;
-import org.hibernate.search.documentation.testsupport.data.ISBN;
 import org.hibernate.search.mapper.orm.Search;
 import org.hibernate.search.mapper.orm.automaticindexing.AutomaticIndexingSynchronizationStrategyName;
 import org.hibernate.search.mapper.orm.cfg.HibernateOrmMapperSettings;
@@ -29,7 +29,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
 @RunWith(Parameterized.class)
-public class ValueBridgeSimpleIT {
+public class PropertyBridgeParameterIT {
 	@Parameterized.Parameters(name = "{0}")
 	public static Object[] backendConfigurations() {
 		return BackendConfigurations.simple().toArray();
@@ -40,7 +40,7 @@ public class ValueBridgeSimpleIT {
 
 	private EntityManagerFactory entityManagerFactory;
 
-	public ValueBridgeSimpleIT(BackendConfiguration backendConfiguration) {
+	public PropertyBridgeParameterIT(BackendConfiguration backendConfiguration) {
 		this.setupHelper = OrmSetupHelper.withSingleBackend( backendConfiguration );
 	}
 
@@ -51,23 +51,34 @@ public class ValueBridgeSimpleIT {
 						HibernateOrmMapperSettings.AUTOMATIC_INDEXING_SYNCHRONIZATION_STRATEGY,
 						AutomaticIndexingSynchronizationStrategyName.SEARCHABLE
 				)
-				.setup( Book.class );
+				.setup( Invoice.class );
 	}
 
 	@Test
 	public void smoke() {
 		OrmUtils.withinJPATransaction( entityManagerFactory, entityManager -> {
-			Book book = new Book();
-			book.setIsbn( ISBN.parse( "978-0-58-600835-5" ) );
-			entityManager.persist( book );
+			Invoice invoice = new Invoice();
+			invoice.getLineItems()
+					.add( new InvoiceLineItem( InvoiceLineItemCategory.BOOK, new BigDecimal( "5.99" ) ) );
+			invoice.getLineItems()
+					.add( new InvoiceLineItem( InvoiceLineItemCategory.BOOK, new BigDecimal( "8.99" ) ) );
+			invoice.getLineItems()
+					.add( new InvoiceLineItem( InvoiceLineItemCategory.BOOK, new BigDecimal( "15.99" ) ) );
+			invoice.getLineItems()
+					.add( new InvoiceLineItem( InvoiceLineItemCategory.SHIPPING, new BigDecimal( "7.99" ) ) );
+			entityManager.persist( invoice );
 		} );
 
 		OrmUtils.withinJPATransaction( entityManagerFactory, entityManager -> {
 			SearchSession searchSession = Search.session( entityManager );
 
-			List<Book> result = searchSession.search( Book.class )
-					.predicate( f -> f.match().field( "isbn" )
-							.matching( ISBN.parse( "978-0-58-600835-5" ) ) )
+			List<Invoice> result = searchSession.search( Invoice.class )
+					.predicate( f -> f.bool()
+							.must( f.range().field( "itemSummary.total" )
+									.atLeast( new BigDecimal( "20.0" ) ) )
+							.must( f.range().field( "itemSummary.shipping" )
+									.atMost( new BigDecimal( "10.0" ) ) )
+					)
 					.fetchHits( 20 );
 
 			assertThat( result ).hasSize( 1 );
